+++ cms/src/routes/__root.tsx
@@ -1,0 +1,12 @@
+import { createRootRoute } from "@tanstack/react-router";
+import App from "../app/App";
+
+const rootRoute = createRootRoute({
+  component: () => (
+    <>
+      <p>hello world</p>
+    </>
+  ),
+});
+
+export default rootRoute;